mudanças na estrutura do banco, melhorias na conta da DA e VC

+++ banco.sql
@@ -13,10 +13,10 @@ CREATE TABLE bem(
 	dt_adiquicao 	DATE NOT NULL,
 	valor_compra 	FLOAT NOT NULL,
     turno 			INT NOT NULL,
-	vida_util 		FLOAT NOT NULL,
+	vida_util 		INT NOT NULL,
 	dt_venda 		DATE,
 	valor_venda 	FLOAT,
-	usado 			FLOAT,
+	usado 			INT,
 	valor_residual 	FLOAT,
 	id_usuario  	INT NOT NULL REFERENCES usuario(id)
 );
